requirements.txt
streamlit==1.32.0
google-generativeai==0.8.4
python-dotenv==1.0.0
